# api-gw/docs/swagger.yaml
basePath: /api
definitions:
  auth.loginReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  auth.registerReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
info:
  contact: {}
  description: REST gateway for the Harmonia project. Orchestrates Python (gRPC) and
    C++ (Thrift) services.
  title: Harmonia API
  version: 0.1.1
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Create a session and set an HTTP-only cookie
      parameters:
      - description: Credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: Invalidate the session and clear the cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the logged-in user if any
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Current user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a user with unique username; logs user in by setting the
        session cookie
      parameters:
      - description: New user
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.registerReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /engine/hello:
    get:
      description: Triggers the Hello RPC on the C++ Thrift EngineService
      parameters:
      - default: World
        description: Name to greet
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Call C++ EngineService Hello RPC
      tags:
      - engine
  /healthz:
    get:
      description: Liveness/readiness probe endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /hello:
    get:
      description: Basic greeting from Harmonia API Gateway
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hello endpoint
      tags:
      - root
  /logic/hello:
    get:
      description: Triggers the Hello RPC on the Python gRPC LogicService
      parameters:
      - default: World
        description: Name to greet
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Call Python LogicService Hello RPC
      tags:
      - logic
swagger: "2.0"
tags:
- description: Root endpoints
  name: root
- description: Authentication endpoints
  name: auth
- description: Python gRPC LogicService
  name: logic
- description: C++ Thrift EngineService
  name: engine
- description: Liveness & readiness
  name: health