# api-gw/internal/engine/handler.go
package engine

import (
	"context"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
)

// HelloEngineRPC godoc
// @Summary      Call C++ EngineService Hello RPC
// @Description  Triggers the Hello RPC on the C++ Thrift EngineService
// @Tags         engine
// @Produce      json
// @Param        name  query  string  false  "Name to greet"  default(World)
// @Success      200   {object}  map[string]string
// @Failure      500   {object}  map[string]string
// @Router       /engine/hello [get]
func Register(rg *gin.RouterGroup, svc *Service) {
	g := rg.Group("/engine")

	g.GET("/hello", func(c *gin.Context) {
		name := c.DefaultQuery("name", "World")
		ctx, cancel := context.WithTimeout(c.Request.Context(), 2*time.Second)
		defer cancel()

		msg, err := svc.Hello(ctx, name)
		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "RPC failed: " + err.Error()})
			return
		}
		c.JSON(http.StatusOK, gin.H{"message": msg})
	})
}